Moving UVs and grid in both x and y... lots of artifacts :(

+++ UnrealMasterThesis/Source/ShaderModels/Private/ShaderModels.cpp
@@ -212,9 +212,9 @@ void ShaderModelsModule::ComputeObstruction(
 	UTextureRenderTarget2D* vPrev_rtt,
 	float xPos,
 	float yPos,
-	float boat_dx,
-	float boat_dy,
-	int speedScale,
+	int boat_dx,
+	int boat_dy,
+	float speedScale,
 	int preFFT) {
 
 	TShaderMapRef<ObstructionShader> shader(GetGlobalShaderMap(GMaxRHIFeatureLevel));
